Split code base by files
